requirements.txt
beautifulsoup4==4.12.2
html2text==2020.1.16
Markdown==3.4.3
Pillow==9.5.0